Reduce duplication in extconf.rb
